rag_app/requirements.txt
python-dotenv
langchain==0.2.1
langchain-community==0.2.1
langchain-core==0.2.7
ollama==0.1.7
psycopg2-binary==2.9.9
tiktoken
datasets
pandas
requests
pgvector
tabulate